requirements.txt
allure_python_commons==2.13.5
chrome_version==0.3.1
pyodbc==5.2.0
pytest==8.3.5
selenium==4.31.0
webdriver_manager==4.0.2
pytest-xdist==3.6.1
